Fix warning about loss of precision.

+++ src/geometry/TankSceneNode.cxx
@@ -731,8 +731,8 @@ void TankIDLSceneNode::IDLRenderNode::render()
   const GLfloat* sphere = sceneNode->tank->getSphere();
   const GLfloat* _plane = sceneNode->plane;
   const GLfloat azimuth = sceneNode->tank->azimuth;
-  const GLfloat ca = cosf(-azimuth * DEG2RAD);
-  const GLfloat sa = sinf(-azimuth * DEG2RAD);
+  const GLfloat ca = cosf(-azimuth * (float)DEG2RAD);
+  const GLfloat sa = sinf(-azimuth * (float)DEG2RAD);
   GLfloat tankPlane[4];
   tankPlane[0] = ca * _plane[0] - sa * _plane[1];
   tankPlane[1] = sa * _plane[0] + ca * _plane[1];
